+++ mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/discount/MLBusinessDiscountBoxView.java
@@ -127,8 +127,8 @@ public class MLBusinessDiscountBoxView extends ConstraintLayout {
         private final int defaultColumns;
 
         SpacesItemDecoration(final Context context, final int itemsInLastRow, final int defaultColumns) {
-            topSpace = (int) ScaleUtils.getPxFromDp(context, 24);
-            lateralSpace = (int) ScaleUtils.getPxFromDp(context, 16);
+            topSpace = (int) ScaleUtils.getPxFromDp(context, 0);
+            lateralSpace = (int) ScaleUtils.getPxFromDp(context, 0);
             this.itemsInLastRow = itemsInLastRow;
             this.defaultColumns = defaultColumns;
         }
@@ -141,7 +141,7 @@ public class MLBusinessDiscountBoxView extends ConstraintLayout {
                 if (itemsInLastRow > 0 &&
                     startLastRow(parent, adapter, view, itemsInLastRow)) {
                     if (itemsInLastRow == 2) {
-                        outRect.left = (view.getLayoutParams().width + lateralSpace * 4) / 2;
+                        setRectLeft(outRect, view, lateralSpace);
                     } else {
                         outRect.left = view.getLayoutParams().width + lateralSpace * 3;
                     }
@@ -157,6 +157,19 @@ public class MLBusinessDiscountBoxView extends ConstraintLayout {
 
                 outRect.right = lateralSpace;
                 outRect.bottom = 0;
+            }
+        }
+
+        /**
+         * @param outRect holds four integer coordinates for a rectangle.
+         * @param view the item view.
+         * @param lateralSpace the lateral space set  previously.
+         */
+        private void setRectLeft(Rect outRect, View view, int lateralSpace) {
+            if (view.getLayoutParams().width <= 0) {
+                outRect.left = (view.getMinimumWidth() + lateralSpace * 4) / 2;
+            } else {
+                outRect.left = (view.getLayoutParams().width + lateralSpace * 4) / 2;
             }
         }