+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## VERSION 1.1.1
 _28_12_2019_
 * REFACTOR - Overlay on discount box image
+* FIX - Loyalty discounts center item ripple effect fix.
 
 ## VERSION 1.1.0
 _05_12_2019_